+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # button-automator - Change Log
 All notable changes to this project will be documented here.
 
-## [1.0.2] - (2021-06-23)
+## [1.1.0] - (2022-07-10)
+- Moving towards button.config.events
+- Support for events where the last button is held
+- Each button press adds another 300ms to listening cycle
 
+## [1.0.2] - (2021-06-23)
 - Reconnect with USB when it goes down